pyproject.toml
[project]
name = "phonofix"
version = "0.2.0"
description = "Multilingual phonetic-similarity replacement engine — a proper-noun substitution tool based on phonetic similarity, supporting ASR/LLM post-processing."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "JonesHong", email = "latte831104@gmail.com" }
]
keywords = [
    "asr",
    "llm",
    "phonetic",
    "chinese",
    "english",
    "text-correction",
    "speech-recognition",
    "nlp",
    "phonofix",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Processing :: Linguistic",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

# 核心依賴 - 預設安裝全部語言支援
# pip install phonofix (等同於 phonofix[all])
dependencies = [
    "python-Levenshtein>=0.12.2",
    "pypinyin>=0.44.0",
    "Pinyin2Hanzi>=0.1.1",
    "hanziconv>=0.3.2",
    "phonemizer>=3.3.0",
    "cutlet>=0.3.0",
    "fugashi>=1.3.0",
    "unidic-lite>=1.0.0",
]

[project.optional-dependencies]
# 中文支援: pip install "phonofix[ch]"
ch = [
    "pypinyin>=0.44.0",
    "Pinyin2Hanzi>=0.1.1",
    "hanziconv>=0.3.2",
]

# 英文支援: pip install "phonofix[en]"
en = [
    "phonemizer>=3.3.0",
]

# 日文支援: pip install "phonofix[ja]"
ja = [
    "cutlet>=0.3.0",
    "fugashi>=1.3.0",
    "unidic-lite>=1.0.0",
]

# 完整安裝 (與預設相同): pip install "phonofix[all]"
all = [
    "pypinyin>=0.44.0",
    "pinyin2hanzi>=0.1.1",
    "hanziconv>=0.3.2",
    "phonemizer>=3.3.0",
    "cutlet>=0.3.0",
    "fugashi>=1.3.0",
    "unidic-lite>=1.0.0",
]

# 開發依賴: pip install "phonofix[dev]"
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/JonesHong/phonofix"
Repository = "https://github.com/JonesHong/phonofix"
Documentation = "https://github.com/JonesHong/phonofix#readme"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/phonofix"]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"
extend-exclude = ["examples", "tools", "references", "scripts", "backups", "dist"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["phonofix"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --tb=short"
pythonpath = ["src"]

[tool.coverage.run]
source = ["src/phonofix"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
